With Debian >11, config.txt and cmdline.txt is in /boot/firmware

+++ install/run_install.sh
@@ -88,6 +88,9 @@ RED='\033[1;31m'
 WHITE='\033[1;37m'
 NOCOLOR='\033[0m'
 
+# What main version is installed
+DEBIAN_VERSION=$(sed 's/\..*//' /etc/debian_version)
+
 # Changes in the variables below (until the ####### delimiter) will be saved
 # into run/torbox.run and used after the installation (we not recommend to
 # change the values until zou precisely know what you are doing)
@@ -1007,13 +1010,23 @@ fi
 # 11. Disabling Bluetooth
 clear
 echo -e "${RED}[+] Step 11: Because of security considerations, we completely disable the Bluetooth functionality${NOCOLOR}"
-if ! grep "# Added by TorBox" /boot/config.txt ; then
-  sudo printf "\n# Added by TorBox\ndtoverlay=disable-bt\n" | sudo tee -a /boot/config.txt
-  sudo systemctl disable hciuart.service
-  sudo systemctl disable bluetooth.service
-  sudo apt-get -y purge bluez
-  sudo apt-get -y autoremove
+if [ "$DEBIAN_VERSION" -gt "11" ]; then
+  if [ -f "/boot/firmware/config.txt" ] ; then
+    if ! grep "# Added by TorBox" /boot/firmware/config.txt ; then
+      sudo printf "\n# Added by TorBox\ndtoverlay=disable-bt\n" | sudo tee -a /boot/firmware/config.txt
+    fi
+  fi
+else
+  if [ -f "/boot/config.txt" ] ; then
+    if ! grep "# Added by TorBox" /boot/config.txt ; then
+      sudo printf "\n# Added by TorBox\ndtoverlay=disable-bt\n" | sudo tee -a /boot/config.txt
+    fi
+  fi
 fi
+sudo systemctl disable hciuart.service
+sudo systemctl disable bluetooth.service
+sudo apt-get -y purge bluez
+sudo apt-get -y autoremove
 sudo rfkill block bluetooth
 
 if [ "$STEP_BY_STEP" = "--step_by_step" ]; then
